backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
psutil==5.9.6
PyJWT==2.8.0
bcrypt==4.1.1
websockets==12.0
